move search model to multiple files
