+++ src/vs/workbench/contrib/search/browser/replace.ts
@@ -3,9 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Match, FileMatch, FileMatchOrMatch } from './searchModel.js';
 import { createDecorator } from '../../../../platform/instantiation/common/instantiation.js';
 import { IProgress, IProgressStep } from '../../../../platform/progress/common/progress.js';
+import { IFileInstanceMatch } from './searchTreeModel/ISearchTreeBase.js';
+import { Match } from './searchTreeModel/match.js';
+import { FileMatchOrMatch } from './searchTreeModel/searchTreeCommon.js';
 
 export const IReplaceService = createDecorator<IReplaceService>('replaceService');
 
@@ -22,7 +24,7 @@ export interface IReplaceService {
 	 *	Replace all the matches from the given file matches in the files
 	 *  You can also pass the progress runner to update the progress of replacing.
 	 */
-	replace(files: FileMatch[], progress?: IProgress<IProgressStep>): Promise<any>;
+	replace(files: IFileInstanceMatch[], progress?: IProgress<IProgressStep>): Promise<any>;
 
 	/**
 	 * Opens the replace preview for given file match or match
@@ -33,5 +35,5 @@ export interface IReplaceService {
 	 * Update the replace preview for the given file.
 	 * If `override` is `true`, then replace preview is constructed from source model
 	 */
-	updateReplacePreview(file: FileMatch, override?: boolean): Promise<void>;
+	updateReplacePreview(file: IFileInstanceMatch, override?: boolean): Promise<void>;
 }
