+++ src/vs/workbench/contrib/search/browser/searchResultsView.ts
@@ -18,7 +18,6 @@ import { ISearchConfigurationProperties } from '../../../services/search/common/
 import { IWorkspaceContextService } from '../../../../platform/workspace/common/workspace.js';
 import { IResourceLabel, ResourceLabels } from '../../../browser/labels.js';
 import { SearchView } from './searchView.js';
-import { FileMatch, Match, RenderableMatch, SearchModel, FolderMatch, FolderMatchNoRoot, FolderMatchWorkspaceRoot, TextSearchResult, AI_TEXT_SEARCH_RESULT_ID } from './searchModel.js';
 import { isEqual } from '../../../../base/common/resources.js';
 import { ICompressibleTreeRenderer } from '../../../../base/browser/ui/tree/objectTree.js';
 import { ICompressedTreeNode } from '../../../../base/browser/ui/tree/compressedObjectTreeModel.js';
@@ -32,6 +31,11 @@ import { defaultCountBadgeStyles } from '../../../../platform/theme/browser/defa
 import { SearchContext } from '../common/constants.js';
 import { getDefaultHoverDelegate } from '../../../../base/browser/ui/hover/hoverDelegateFactory.js';
 import { IHoverService } from '../../../../platform/hover/browser/hover.js';
+import { FolderMatchImpl, FolderMatchWorkspaceRootImpl, FolderMatchNoRootImpl } from './searchTreeModel/folderMatch.js';
+import { Match } from './searchTreeModel/match.js';
+import { SearchModelImpl } from './searchTreeModel/searchModel.js';
+import { RenderableMatch, AI_TEXT_SEARCH_RESULT_ID } from './searchTreeModel/searchTreeCommon.js';
+import { ITextSearchHeading, IFolderMatch, IFileInstanceMatch, isFileInstanceMatch, isFolderMatch, isTextSearchHeading } from './searchTreeModel/ISearchTreeBase.js';
 
 interface IFolderMatchTemplate {
 	label: IResourceLabel;
@@ -78,13 +82,13 @@ export class SearchDelegate implements IListVirtualDelegate<RenderableMatch> {
 	}
 
 	getTemplateId(element: RenderableMatch): string {
-		if (element instanceof FolderMatch) {
+		if (isFolderMatch(element)) {
 			return FolderMatchRenderer.TEMPLATE_ID;
-		} else if (element instanceof FileMatch) {
+		} else if (isFileInstanceMatch(element)) {
 			return FileMatchRenderer.TEMPLATE_ID;
 		} else if (element instanceof Match) {
 			return MatchRenderer.TEMPLATE_ID;
-		} else if (element instanceof TextSearchResult) {
+		} else if (isTextSearchHeading(element)) {
 			return TextSearchResultRenderer.TEMPLATE_ID;
 		}
 
@@ -93,7 +97,7 @@ export class SearchDelegate implements IListVirtualDelegate<RenderableMatch> {
 	}
 }
 
-export class TextSearchResultRenderer extends Disposable implements ICompressibleTreeRenderer<TextSearchResult, any, ITextSearchResultTemplate> {
+export class TextSearchResultRenderer extends Disposable implements ICompressibleTreeRenderer<ITextSearchHeading, any, ITextSearchResultTemplate> {
 	static readonly TEMPLATE_ID = 'textResultMatch';
 
 	readonly templateId = TextSearchResultRenderer.TEMPLATE_ID;
@@ -104,7 +108,7 @@ export class TextSearchResultRenderer extends Disposable implements ICompressibl
 	) {
 		super();
 	}
-	disposeCompressedElements?(node: ITreeNode<ICompressedTreeNode<TextSearchResult>, any>, index: number, templateData: ITextSearchResultTemplate, height: number | undefined): void {
+	disposeCompressedElements?(node: ITreeNode<ICompressedTreeNode<ITextSearchHeading>, any>, index: number, templateData: ITextSearchResultTemplate, height: number | undefined): void {
 
 	}
 	renderTemplate(container: HTMLElement): ITextSearchResultTemplate {
@@ -115,7 +119,7 @@ export class TextSearchResultRenderer extends Disposable implements ICompressibl
 		return { label, disposables };
 	}
 
-	async renderElement(node: ITreeNode<TextSearchResult, any>, index: number, templateData: IFolderMatchTemplate, height: number | undefined): Promise<void> {
+	async renderElement(node: ITreeNode<ITextSearchHeading, any>, index: number, templateData: IFolderMatchTemplate, height: number | undefined): Promise<void> {
 		if (node.element.id() === AI_TEXT_SEARCH_RESULT_ID) {
 			const aiName = await node.element.parent().searchModel.getAITextResultProviderName();
 			templateData.label.setLabel(nls.localize({
@@ -131,11 +135,11 @@ export class TextSearchResultRenderer extends Disposable implements ICompressibl
 		templateData.disposables.dispose();
 	}
 
-	renderCompressedElements(node: ITreeNode<ICompressedTreeNode<TextSearchResult>, any>, index: number, templateData: ITextSearchResultTemplate, height: number | undefined): void {
+	renderCompressedElements(node: ITreeNode<ICompressedTreeNode<ITextSearchHeading>, any>, index: number, templateData: ITextSearchResultTemplate, height: number | undefined): void {
 	}
 
 }
-export class FolderMatchRenderer extends Disposable implements ICompressibleTreeRenderer<FolderMatch, any, IFolderMatchTemplate> {
+export class FolderMatchRenderer extends Disposable implements ICompressibleTreeRenderer<IFolderMatch, any, IFolderMatchTemplate> {
 	static readonly TEMPLATE_ID = 'folderMatch';
 
 	readonly templateId = FolderMatchRenderer.TEMPLATE_ID;
@@ -151,13 +155,13 @@ export class FolderMatchRenderer extends Disposable implements ICompressibleTree
 		super();
 	}
 
-	renderCompressedElements(node: ITreeNode<ICompressedTreeNode<FolderMatch>, any>, index: number, templateData: IFolderMatchTemplate, height: number | undefined): void {
+	renderCompressedElements(node: ITreeNode<ICompressedTreeNode<FolderMatchImpl>, any>, index: number, templateData: IFolderMatchTemplate, height: number | undefined): void {
 		const compressed = node.element;
 		const folder = compressed.elements[compressed.elements.length - 1];
 		const label = compressed.elements.map(e => e.name());
 
 		if (folder.resource) {
-			const fileKind = (folder instanceof FolderMatchWorkspaceRoot) ? FileKind.ROOT_FOLDER : FileKind.FOLDER;
+			const fileKind = (folder instanceof FolderMatchWorkspaceRootImpl) ? FileKind.ROOT_FOLDER : FileKind.FOLDER;
 			templateData.label.setResource({ resource: folder.resource, name: label }, {
 				fileKind,
 				separator: this.labelService.getSeparator(folder.resource.scheme),
@@ -207,7 +211,7 @@ export class FolderMatchRenderer extends Disposable implements ICompressibleTree
 		};
 	}
 
-	renderElement(node: ITreeNode<FolderMatch, any>, index: number, templateData: IFolderMatchTemplate): void {
+	renderElement(node: ITreeNode<FolderMatchImpl, any>, index: number, templateData: IFolderMatchTemplate): void {
 		const folderMatch = node.element;
 		if (folderMatch.resource) {
 			const workspaceFolder = this.contextService.getWorkspaceFolder(folderMatch.resource);
@@ -233,7 +237,7 @@ export class FolderMatchRenderer extends Disposable implements ICompressibleTree
 		templateData.elementDisposables.clear();
 	}
 
-	disposeCompressedElements(node: ITreeNode<ICompressedTreeNode<FolderMatch>, any>, index: number, templateData: IFolderMatchTemplate, height: number | undefined): void {
+	disposeCompressedElements(node: ITreeNode<ICompressedTreeNode<FolderMatchImpl>, any>, index: number, templateData: IFolderMatchTemplate, height: number | undefined): void {
 		templateData.elementDisposables.clear();
 	}
 
@@ -241,7 +245,7 @@ export class FolderMatchRenderer extends Disposable implements ICompressibleTree
 		templateData.disposables.dispose();
 	}
 
-	private renderFolderDetails(folder: FolderMatch, templateData: IFolderMatchTemplate) {
+	private renderFolderDetails(folder: FolderMatchImpl, templateData: IFolderMatchTemplate) {
 		const count = folder.recursiveMatchCount();
 		templateData.badge.setCount(count);
 		templateData.badge.setTitleFormat(count > 1 ? nls.localize('searchFileMatches', "{0} files found", count) : nls.localize('searchFileMatch', "{0} file found", count));
@@ -250,7 +254,7 @@ export class FolderMatchRenderer extends Disposable implements ICompressibleTree
 	}
 }
 
-export class FileMatchRenderer extends Disposable implements ICompressibleTreeRenderer<FileMatch, any, IFileMatchTemplate> {
+export class FileMatchRenderer extends Disposable implements ICompressibleTreeRenderer<IFileInstanceMatch, any, IFileMatchTemplate> {
 	static readonly TEMPLATE_ID = 'fileMatch';
 
 	readonly templateId = FileMatchRenderer.TEMPLATE_ID;
@@ -266,7 +270,7 @@ export class FileMatchRenderer extends Disposable implements ICompressibleTreeRe
 		super();
 	}
 
-	renderCompressedElements(node: ITreeNode<ICompressedTreeNode<FileMatch>, any>, index: number, templateData: IFileMatchTemplate, height: number | undefined): void {
+	renderCompressedElements(node: ITreeNode<ICompressedTreeNode<IFileInstanceMatch>, any>, index: number, templateData: IFileMatchTemplate, height: number | undefined): void {
 		throw new Error('Should never happen since node is incompressible.');
 	}
 
@@ -307,12 +311,12 @@ export class FileMatchRenderer extends Disposable implements ICompressibleTreeRe
 		};
 	}
 
-	renderElement(node: ITreeNode<FileMatch, any>, index: number, templateData: IFileMatchTemplate): void {
+	renderElement(node: ITreeNode<IFileInstanceMatch, any>, index: number, templateData: IFileMatchTemplate): void {
 		const fileMatch = node.element;
 		templateData.el.setAttribute('data-resource', fileMatch.resource.toString());
 
 		const decorationConfig = this.configurationService.getValue<ISearchConfigurationProperties>('search').decorations;
-		templateData.label.setFile(fileMatch.resource, { hidePath: this.searchView.isTreeLayoutViewVisible && !(fileMatch.parent() instanceof FolderMatchNoRoot), hideIcon: false, fileDecorations: { colors: decorationConfig.colors, badges: decorationConfig.badges } });
+		templateData.label.setFile(fileMatch.resource, { hidePath: this.searchView.isTreeLayoutViewVisible && !(fileMatch.parent() instanceof FolderMatchNoRootImpl), hideIcon: false, fileDecorations: { colors: decorationConfig.colors, badges: decorationConfig.badges } });
 		const count = fileMatch.count();
 		templateData.badge.setCount(count);
 		templateData.badge.setTitleFormat(count > 1 ? nls.localize('searchMatches', "{0} matches found", count) : nls.localize('searchMatch', "{0} match found", count));
@@ -466,14 +470,14 @@ export class SearchAccessibilityProvider implements IListAccessibilityProvider<R
 	}
 
 	getAriaLabel(element: RenderableMatch): string | null {
-		if (element instanceof FolderMatch) {
+		if (isFolderMatch(element)) {
 			const count = element.allDownstreamFileMatches().reduce((total, current) => total + current.count(), 0);
 			return element.resource ?
 				nls.localize('folderMatchAriaLabel', "{0} matches in folder root {1}, Search result", count, element.name()) :
 				nls.localize('otherFilesAriaLabel', "{0} matches outside of the workspace, Search result", count);
 		}
 
-		if (element instanceof FileMatch) {
+		if (isFileInstanceMatch(element)) {
 			const path = this.labelService.getUriLabel(element.resource, { relative: true }) || element.resource.fsPath;
 
 			return nls.localize('fileMatchAriaLabel', "{0} matches in file {1} of folder {2}, Search result", element.count(), element.name(), paths.dirname(path));
@@ -481,7 +485,7 @@ export class SearchAccessibilityProvider implements IListAccessibilityProvider<R
 
 		if (element instanceof Match) {
 			const match = <Match>element;
-			const searchModel: SearchModel = this.searchView.model;
+			const searchModel: SearchModelImpl = this.searchView.model;
 			const replace = searchModel.isReplaceActive() && !!searchModel.replaceString;
 			const matchString = match.getMatchString();
 			const range = match.range();
