+++ src/vs/workbench/contrib/search/browser/searchTreeModel/match.ts
@@ -1,0 +1,146 @@
+import { memoize } from '../../../../../base/common/decorators';
+import { lcut } from '../../../../../base/common/strings';
+import { ISearchRange, OneLineRange } from '../../../../services/search/common/search';
+import { Range } from '../../../../../editor/common/core/range.js';
+import { IFileInstanceMatch } from './ISearchTreeBase';
+
+export class Match {
+
+	private static readonly MAX_PREVIEW_CHARS = 250;
+	protected _id: string;
+	protected _range: Range;
+	private _oneLinePreviewText: string;
+	private _rangeInPreviewText: ISearchRange;
+	// For replace
+	private _fullPreviewRange: ISearchRange;
+
+	constructor(protected _parent: IFileInstanceMatch, private _fullPreviewLines: string[], _fullPreviewRange: ISearchRange, _documentRange: ISearchRange, public readonly aiContributed: boolean) {
+		this._oneLinePreviewText = _fullPreviewLines[_fullPreviewRange.startLineNumber];
+		const adjustedEndCol = _fullPreviewRange.startLineNumber === _fullPreviewRange.endLineNumber ?
+			_fullPreviewRange.endColumn :
+			this._oneLinePreviewText.length;
+		this._rangeInPreviewText = new OneLineRange(1, _fullPreviewRange.startColumn + 1, adjustedEndCol + 1);
+
+		this._range = new Range(
+			_documentRange.startLineNumber + 1,
+			_documentRange.startColumn + 1,
+			_documentRange.endLineNumber + 1,
+			_documentRange.endColumn + 1);
+
+		this._fullPreviewRange = _fullPreviewRange;
+
+		this._id = this._parent.id() + '>' + this._range + this.getMatchString();
+	}
+
+	id(): string {
+		return this._id;
+	}
+
+	parent(): IFileInstanceMatch {
+		return this._parent;
+	}
+
+	text(): string {
+		return this._oneLinePreviewText;
+	}
+
+	range(): Range {
+		return this._range;
+	}
+
+	@memoize
+	preview(): { before: string; fullBefore: string; inside: string; after: string } {
+		const fullBefore = this._oneLinePreviewText.substring(0, this._rangeInPreviewText.startColumn - 1),
+			before = lcut(fullBefore, 26, '…');
+
+		let inside = this.getMatchString(),
+			after = this._oneLinePreviewText.substring(this._rangeInPreviewText.endColumn - 1);
+
+		let charsRemaining = Match.MAX_PREVIEW_CHARS - before.length;
+		inside = inside.substr(0, charsRemaining);
+		charsRemaining -= inside.length;
+		after = after.substr(0, charsRemaining);
+
+		return {
+			before,
+			fullBefore,
+			inside,
+			after,
+		};
+	}
+
+	get replaceString(): string {
+		const searchModel = this.parent().parent().searchModel;
+		if (!searchModel.replacePattern) {
+			throw new Error('searchModel.replacePattern must be set before accessing replaceString');
+		}
+
+		const fullMatchText = this.fullMatchText();
+		let replaceString = searchModel.replacePattern.getReplaceString(fullMatchText, searchModel.preserveCase);
+		if (replaceString !== null) {
+			return replaceString;
+		}
+
+		// Search/find normalize line endings - check whether \r prevents regex from matching
+		const fullMatchTextWithoutCR = fullMatchText.replace(/\r\n/g, '\n');
+		if (fullMatchTextWithoutCR !== fullMatchText) {
+			replaceString = searchModel.replacePattern.getReplaceString(fullMatchTextWithoutCR, searchModel.preserveCase);
+			if (replaceString !== null) {
+				return replaceString;
+			}
+		}
+
+		// If match string is not matching then regex pattern has a lookahead expression
+		const contextMatchTextWithSurroundingContent = this.fullMatchText(true);
+		replaceString = searchModel.replacePattern.getReplaceString(contextMatchTextWithSurroundingContent, searchModel.preserveCase);
+		if (replaceString !== null) {
+			return replaceString;
+		}
+
+		// Search/find normalize line endings, this time in full context
+		const contextMatchTextWithoutCR = contextMatchTextWithSurroundingContent.replace(/\r\n/g, '\n');
+		if (contextMatchTextWithoutCR !== contextMatchTextWithSurroundingContent) {
+			replaceString = searchModel.replacePattern.getReplaceString(contextMatchTextWithoutCR, searchModel.preserveCase);
+			if (replaceString !== null) {
+				return replaceString;
+			}
+		}
+
+		// Match string is still not matching. Could be unsupported matches (multi-line).
+		return searchModel.replacePattern.pattern;
+	}
+
+	fullMatchText(includeSurrounding = false): string {
+		let thisMatchPreviewLines: string[];
+		if (includeSurrounding) {
+			thisMatchPreviewLines = this._fullPreviewLines;
+		} else {
+			thisMatchPreviewLines = this._fullPreviewLines.slice(this._fullPreviewRange.startLineNumber, this._fullPreviewRange.endLineNumber + 1);
+			thisMatchPreviewLines[thisMatchPreviewLines.length - 1] = thisMatchPreviewLines[thisMatchPreviewLines.length - 1].slice(0, this._fullPreviewRange.endColumn);
+			thisMatchPreviewLines[0] = thisMatchPreviewLines[0].slice(this._fullPreviewRange.startColumn);
+		}
+
+		return thisMatchPreviewLines.join('\n');
+	}
+
+	rangeInPreview() {
+		// convert to editor's base 1 positions.
+		return {
+			...this._fullPreviewRange,
+			startColumn: this._fullPreviewRange.startColumn + 1,
+			endColumn: this._fullPreviewRange.endColumn + 1
+		};
+	}
+
+	fullPreviewLines(): string[] {
+		return this._fullPreviewLines.slice(this._fullPreviewRange.startLineNumber, this._fullPreviewRange.endLineNumber + 1);
+	}
+
+	getMatchString(): string {
+		return this._oneLinePreviewText.substring(this._rangeInPreviewText.startColumn - 1, this._rangeInPreviewText.endColumn - 1);
+	}
+
+	public isReadonly() {
+		return this.aiContributed;
+	}
+}
