+++ src/vs/workbench/contrib/search/test/browser/mockSearchTree.ts
@@ -6,7 +6,7 @@
 import { ITreeNavigator } from '../../../../../base/browser/ui/tree/tree.js';
 import { Emitter } from '../../../../../base/common/event.js';
 import { IDisposable } from '../../../../../base/common/lifecycle.js';
-import { RenderableMatch } from '../../browser/searchModel.js';
+import { RenderableMatch } from '../../browser/searchTreeModel/searchModel.js';
 
 const someEvent = new Emitter().event;
